// app/profile/model/person.ts
export class Person {
    Id: number;
    Name: string;
    LastNames: string;
    DateOfBirth: string;
    Sex: boolean;
    Race: string;
    Phone: string;
    Mobile: string;
    Email: string;
    UserId: string;
    DNI: string;
    Active: boolean;
    Age: number;
    SpecialtyId: number;
    strDateOfBirth: string;
    Role: string;
    Image: string;
}